I'll rename the .js files to .jsx.
