+++ src/components/ReceiptUpload.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useContext } from 'react';
 import { AuthContext } from '../App';
 import { CurrencyContext } from '../CurrencyContext';
-import { getRequests, getReceipts, saveReceipt, sendEmailNotification, getUsers } from '../data/models';
+import { getRequests, getReceipts, saveReceipt, sendEmailNotification, getUsers } from '../data/models.jsx';
 
 const ReceiptUpload = () => {
   const { user } = useContext(AuthContext);
