+++ src/main.jsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App.jsx'
 import './index.css'
-import { CurrencyProvider } from './CurrencyContext.js'
+import { CurrencyProvider } from './CurrencyContext.jsx'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
